Typescriptify the other CRUD hooks
