+++ packages/lesswrong/lib/crud/withDelete.tsx
