+++ packages/lesswrong/lib/crud/withMutation.ts
@@ -14,6 +14,7 @@ export default withMutation({
 import { graphql } from 'react-apollo';
 import gql from 'graphql-tag';
 import { getFragment } from 'meteor/vulcan:lib';
+import * as _ from 'underscore';
 
 export function withMutation({name, args, fragmentName}) {
 
@@ -44,7 +45,7 @@ export function withMutation({name, args, fragmentName}) {
   
   return graphql(gql`${mutation}${fragmentName ? fragment : ''}`, {
     alias: 'withMutation',
-    props: ({ownProps, mutate}) => ({
+    props: ({ownProps, mutate}: any) => ({
       [name]: (vars) => {
         return mutate({ 
           variables: vars,
